There is no Arquillian adapter for tomee and jakara 9 yet so we must hemove the test folder

+++ rest-starter/src/main/resources/META-INF/archetype-post-generate.groovy
@@ -14,9 +14,7 @@ switch (request.properties["runtime"])
                       break
                       
     case "tomee":     println "Generating code for TomEE"
-                      if ( jakartaVersion != '8') {
-                          FileUtils.deleteDirectory(new File(outputDirectory, "src/test"))
-                      }
+                      if ( jakartaVersion != '8') FileUtils.deleteDirectory(new File(outputDirectory, "src/test"))
                       break
 
     default:          println "Generating code for Payara"
